+++ lib/main.dart
@@ -13,7 +13,6 @@ void main() async {
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     String? usersLogin = AppSettings.getLogin();
